// apps/developer-app/src/app/models/matrix.ts
export interface Matrix {
  id: number;
  GSSN_OUTLETID: string;
  GSSN_COMPANYID: string;
  vd_name: string;
  vfnr: string;
  vfnr_hb: string;
  mode: string;
  sortorder: string;
  brand: string;
  status: string;
  shortcut: string;
  company_name: string;
  branch: string;
  street: string;
  zip: string;
  place: string;
  region: string;
  phone: string;
  website: string;
  center_name: string;
}